+++ src/extension.ts
@@ -1,7 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
-import { AzureDevopsProvider, Repos } from './providers/azureDevopsProvider';
+import { AzureDevopsProvider, Repos, TestCase, PullRequests } from './providers/azureDevopsProvider';
 import { MicrosoftProviderId, MicrosoftScopes } from './Utils/Constants';
 
 
@@ -14,6 +14,7 @@ export async function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(treeView);
 	context.subscriptions.push(vscode.commands.registerCommand('devops-toolkit.clone', async (item: Repos) => { vscode.commands.executeCommand('git.clone', item.webUrl); }));
 	context.subscriptions.push(vscode.commands.registerCommand('devops-toolkit.refresh', async (item) => { azureDevopsProvider.refresh(item); }));
+	context.subscriptions.push(vscode.commands.registerCommand('devops-toolkit.openInBrowser', async (item: PullRequests | TestCase) => { vscode.env.openExternal(vscode.Uri.parse(item.webUrl)); }));
 }
 
 // This method is called when your extension is deactivated